src/ai/pyproject.toml
[project]
name = "practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.5",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.13.0",
    "attrs==26.1.0",
    "bcrypt==5.0.0",
    "build==1.4.3",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.7",
    "chromadb==1.5.7",
    "click==8.3.2",
    "colorama==0.4.6",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "durationpy==0.10",
    "fastapi==0.135.3",
    "filelock==3.25.2",
    "flatbuffers==25.12.19",
    "frozenlist==1.8.0",
    "fsspec==2026.3.0",
    "google-api-python-client>=2.194.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "googleapis-common-protos==1.74.0",
    "greenlet==3.4.0",
    "grpcio==1.80.0",
    "h11==0.16.0",
    "hf-xet==1.4.3",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==1.10.1",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "importlib-resources==7.1.0",
    "jiter==0.14.0",
    "jsonpatch==1.33",
    "jsonpointer==3.1.1",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "kubernetes==35.0.0",
    "langchain==1.2.15",
    "langchain-chroma==1.1.0",
    "langchain-classic==1.0.3",
    "langchain-community==0.4.1",
    "langchain-core==1.2.28",
    "langchain-openai==1.1.12",
    "langchain-tavily>=0.2.18",
    "langchain-text-splitters==1.1.1",
    "langgraph==1.1.6",
    "langgraph-checkpoint==4.0.1",
    "langgraph-prebuilt==1.0.9",
    "langgraph-sdk==0.3.13",
    "langsmith==0.7.30",
    "markdown-it-py==4.0.0",
    "marshmallow==3.26.2",
    "mdurl==0.1.2",
    "mmh3==5.2.1",
    "mpmath==1.3.0",
    "multidict==6.7.1",
    "mypy-extensions==1.1.0",
    "numpy==2.4.4",
    "oauthlib==3.3.1",
    "onnxruntime==1.24.4",
    "openai==2.31.0",
    "opentelemetry-api==1.41.0",
    "opentelemetry-exporter-otlp-proto-common==1.41.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.41.0",
    "opentelemetry-proto==1.41.0",
    "opentelemetry-sdk==1.41.0",
    "opentelemetry-semantic-conventions==0.62b0",
    "orjson==3.11.8",
    "ormsgpack==1.12.2",
    "overrides==7.7.0",
    "packaging==26.0",
    "propcache==0.4.1",
    "protobuf==6.33.6",
    "pybase64==1.4.3",
    "pydantic==2.13.0",
    "pydantic-core==2.46.0",
    "pydantic-settings==2.13.1",
    "pygments==2.20.0",
    "pypdf==6.10.0",
    "pypika==0.51.1",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.2",
    "python-multipart==0.0.26",
    "pyyaml==6.0.3",
    "referencing==0.37.0",
    "regex==2026.4.4",
    "requests==2.33.1",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==15.0.0",
    "rpds-py==0.30.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.49",
    "starlette==1.0.0",
    "sympy==1.14.0",
    "tavily>=1.1.0",
    "tenacity==9.1.4",
    "tiktoken==0.12.0",
    "tokenizers==0.22.2",
    "tqdm==4.67.3",
    "typer==0.24.1",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uuid-utils==0.14.1",
    "uvicorn==0.44.0",
    "watchfiles==1.1.1",
    "websocket-client==1.9.0",
    "websockets==16.0",
    "xxhash==3.6.0",
    "yarl==1.23.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]
